backend/requirements-dev.txt
-r requirements.txt

# Development
django-debug-toolbar>=4.2,<5.0
ipython>=8.0,<9.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code Quality
ruff>=0.2,<1.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Pre-commit
pre-commit>=3.6,<4.0
